# backend/docs/swagger.yaml
definitions:
...
  dto.SaleDTO:
    properties:
      datetime:
        type: string
      discount:
        type: number
      full_cost:
        type: number
      sale_id:
        type: string
      total_cost:
        type: number
    type: object
...
  dto.SupplierDTO:
    properties:
      email:
        type: string
      inn:
        type: string
      kpp:
        type: string
      name:
        type: string
      ogrn:
        type: string
      phone:
        type: string
      supplier_id:
        type: string
    type: object
...
  models.Sale:
    properties:
      datetime:
        type: string
      discount:
        type: number
      full_cost:
        type: number
      sale_id:
        type: string
      total_cost:
        type: number
    type: object
  models.Supplier:
    properties:
      email:
        type: string
      inn:
        type: string
      kpp:
        type: string
      name:
        type: string
      ogrn:
        type: string
      phone:
        type: string
      supplier_id:
        type: string
    type: object
  response.Response:
    properties:
      error:
        description: Error if response is error
        type: string
      status:
        description: Response status
        type: string
    type: object
info:
  contact: {}
paths:
...
  /sales:
    get:
      description: Returns a list of sales filtered by date range.
      parameters:
      - description: Start date (RFC3339)
        in: query
        name: date_from
        type: string
      - description: End date (RFC3339)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SaleDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get sales
      tags:
      - sales
    post:
      consumes:
      - application/json
      description: Creates a new sale.
      parameters:
      - description: Sale
        in: body
        name: sale
        required: true
        schema:
          $ref: '#/definitions/models.Sale'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Sale'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create sale
      tags:
      - sales
  /sales/{id}:
    delete:
      description: Deletes a sale by ID.
      parameters:
      - description: Sale ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete sale
      tags:
      - sales
    put:
      consumes:
      - application/json
      description: Updates an existing sale.
      parameters:
      - description: Sale ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Sale
        in: body
        name: sale
        required: true
        schema:
          $ref: '#/definitions/models.Sale'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update sale
      tags:
      - sales
...
  /suppliers:
    delete:
      description: Deletes a supplier by ID.
      parameters:
      - description: Supplier ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Supplier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete supplier
      tags:
      - suppliers
    get:
      description: Returns a list of suppliers filtered by search string.
      parameters:
      - description: Search string
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SupplierDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get suppliers
      tags:
      - suppliers
    post:
      consumes:
      - application/json
      description: Creates a new supplier.
      parameters:
      - description: Supplier
        in: body
        name: supplier
        required: true
        schema:
          $ref: '#/definitions/models.Supplier'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Supplier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create supplier
      tags:
      - suppliers
    put:
      consumes:
      - application/json
      description: Updates a supplier by ID.
      parameters:
      - description: Supplier ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Supplier
        in: body
        name: supplier
        required: true
        schema:
          $ref: '#/definitions/models.Supplier'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Supplier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update supplier
      tags:
      - suppliers
swagger: "2.0"